refactor: migrate getByAPI & queryByAPI
